size_t->int64_t + removed unused var in load graph

+++ cpp/oneapi/dal/algo/jaccard/backend/cpu/vertex_similarity_default_kernel_avx2.hpp
@@ -48,11 +48,11 @@ DAAL_FORCEINLINE std::int32_t _popcnt32_redef(const std::int32_t &x) {
 }
 #endif
 
-DAAL_FORCEINLINE std::size_t intersection(const std::int32_t *neigh_u,
+DAAL_FORCEINLINE std::int64_t intersection(const std::int32_t *neigh_u,
                                           const std::int32_t *neigh_v,
                                           std::int32_t n_u,
                                           std::int32_t n_v) {
-    std::size_t total = 0;
+    std::int64_t total = 0;
     std::int32_t i_u = 0, i_v = 0;
 
     const std::int32_t n_u_8_end = n_u - 8;
